backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.16

# Data processing
numpy==1.26.0
opencv-python-headless==4.9.0.80
pillow==10.2.0

# AI/ML - CPU only for Render free tier
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# GradCAM Visualization
pytorch-grad-cam==1.5.0

# ONNX Runtime (lighter than full PyTorch)
onnxruntime==1.16.0

# Other
pydantic==2.5.0
pydantic-settings==2.1.0
